@@ -94,7 +94,7 @@ def init_surface_recon_wf(
     Second, the brain mask is grown, including pixels that have a high likelihood
     to the GM tissue distribution:
 
-      3. Dilate and substract the brain mask, defining the region to search for candidate
+      3. Dilate and subtract the brain mask, defining the region to search for candidate
          pixels that likely belong to cortical GM.
 
       4. Pixels found in the search region that are labeled as GM by ANTs
@@ -520,7 +520,7 @@ def init_autorecon_resume_wf(*, omp_nthreads, name="autorecon_resume_wf"):
         vals = set(in_list)
         if len(vals) > 1:
             raise ValueError(
-                "Non-identical values can't be deduplicated:\n{!r}".format(in_list)
+                f"Non-identical values can't be deduplicated:\n{in_list!r}"
             )
         return vals.pop()
 
@@ -750,7 +750,7 @@ def init_gifti_surface_wf(*, name="gifti_surface_wf"):
     If midthickness (or graymid) surfaces do not exist, they are generated and
     saved to the subject directory as ``lh/rh.midthickness``.
     These, along with the gray/white matter boundary (``lh/rh.white``), pial
-    sufaces (``lh/rh.pial``) and inflated surfaces (``lh/rh.inflated``) are
+    surfaces (``lh/rh.pial``) and inflated surfaces (``lh/rh.inflated``) are
     converted to GIFTI files.
     Additionally, the vertex coordinates are :py:class:`recentered
     <smriprep.interfaces.NormalizeSurf>` to align with native T1w space.
@@ -886,7 +886,7 @@ def init_segs_to_native_wf(*, name="segs_to_native", segmentation="aseg"):
         The selected segmentation, after resampling in native space
 
     """
-    workflow = Workflow(name="%s_%s" % (name, segmentation))
+    workflow = Workflow(name=f"{name}_{segmentation}")
     inputnode = pe.Node(
         niu.IdentityInterface(["in_file", "subjects_dir", "subject_id", "fsnative2t1w_xfm"]),
         name="inputnode",
