@@ -24,10 +24,7 @@
 Base module variables
 """
 
-from ._version import get_versions
-
-__version__ = get_versions()["version"]
-del get_versions
+from ._version import __version__
 
 __copyright__ = (
     "Copyright 2019, Center for Reproducible Neuroscience, Stanford University"
