@@ -980,8 +980,8 @@ def init_anat_second_derivatives_wf(
         (inputnode, ds_regs, [('sphere_reg', 'in_file'),
                               ('source_files', 'source_file')]),
         (name_regs, ds_regs, [('hemi', 'hemi')]),
-        (inputnode, name_reg_fsLR, [('sphere_reg', 'in_file')]),
-        (inputnode, ds_reg_fsLR, [('sphere_reg', 'in_file'),
+        (inputnode, name_reg_fsLR, [('sphere_reg_fsLR', 'in_file')]),
+        (inputnode, ds_reg_fsLR, [('sphere_reg_fsLR', 'in_file'),
                                   ('source_files', 'source_file')]),
         (name_reg_fsLR, ds_reg_fsLR, [('hemi', 'hemi')]),
         (inputnode, name_morphs, [('morphometrics', 'in_file')]),
